about the Cocoon-Centers-Project
Bootcamp software is a nice camp organization for those who know a degree. But you want us to have video tutorials for those who don't know or have little knowledge. Those who do not have a suitable environment for work or who cannot find a calm, quiet and non-stressful environment to work at home, those who cannot work or focus due to the stress of the home or the comfortable environment, those who do not live in cities such as Istanbul where there are many opportunities, those who do not have computer facilities, those who do not have the software offered by companies in specially reserved areas for working. people who have the problem of not having a suitable area to work in provinces that do not have a suitable working environment for bla bla, in a quiet, stress-free environment, with free coffee, tea and internet freedom, where computer desks are offered for those who do not have a laptop, and cocoon areas sponsored by companies or foundations I want her to leave. Beginning programmer candidates with larvae in cocoons will complete their education online and embark on a journey to become a junior software developer. I introduce the name of this project concept as bitcamp. Candidates who train themselves in Bitcamp cocoon centers will be able to switch to bootcamp centers. This transition will be possible. For those who prefer to stay in Bitcamps, the freedom to develop themselves and earn income by doing freelance work when they become Juniors will be offered at bitcamp cocoon centers. People who have completed their Bitcamp journey and entered the active job market will now be able to become butterflies and leave the nest. Bitcamp cocoon centers should also cooperate with Entrepreneurship incubation centers. Butterflies that will fly from our Cocoon Centers will be able to become entrepreneurs and establish start-ups in incubation centers if they wish. In these start-ups, cocoon center areas where our butterflies can leave and raise their larvae will also be reserved. Each start-up will also be a cocoon center. The larvae participating in the bitcamps will raise themselves in the cocoon centers located within the domains of these start-up companies. Later, if they wish, when they become juniors, they will start working for the start-up they are in, or within the cocoon start-up companies in any other place or city they wish, contributing to the start-ups and going to the pupa stage. Then, our senior junior software developer will leave the pupa stage and move to important positions in the job market, flying as a butterfly software developer and flapping his wings towards the place where he will build his own nest. In this way, we will build a sustainable and expandable system. It will be sustainable because the growing butterfly software developers will raise their own larvae in the start-ups they will establish. Thus, the system will find a solution for its own workforce needs. It will be expandable because more and more butterfly software developers will grow and the circle will expand by multiplying the number of cocoon centers that will allow these developers to grow more larvae. With the bitcamps that these cocoon centers will organize, they will have free environments where they can raise themselves for young people whose environment and situation are not suitable. These opportunities will be provided all over Anatolia. Those admitted to the bitcamps will be selected through interviews based solely on cultural tests and a love of software. Bigots and blinders will not be accepted, whose worldview does not have a wide horizon. The aim should be to choose people who will serve humanity, not a group, and pave the way for them. The second interview test, the software love test, is to measure their sincere feelings about the software and to understand whether they really want to be a software developer. In this way, the larvae participating in the bitcamps will be individuals that are raised for humanity and civilization. bitcamps will fill that gap. Bootcamps offer an infrastructure that trains software developer candidates who have reached a level in terms of knowledge, but who have no experience, and enable them to be appointed as juniors to companies. If other companies are not going to buy from bootcamps, they are at least interested in junior software developers. Companies with a need for senior software hire Senior software developers. There is no system, organization or camp that will provide an opportunity for improvement between the zero level and the level of Junior candidates who do not have certain knowledge and experience. bitcamps will provide a System and Infrastructure that will fill the gap at the lowest level in the whole, which we can call the first level with this zero. Climbing this first step is a very difficult situation for millions of individuals who want to become software developers. They cannot really work at home, because they are overwhelmed by the pressure of finding a job, helping with the housework, pressure from the neighborhood, etc. and they cannot focus because they do not feel free. Because they are scattered, they cannot focus on improving themselves, and this triggers uncertainty and future anxiety in them. Anxiety problems make it more difficult for them to focus. And people who are or feel surrounded by chains find themselves in a vicious circle that is difficult to break. Some of them do not have financial means, do not have a computer, do not have internet. Some also need a friend, a friend to help them stay motivated. Here, the cocoon centers will break this chain, break the vicious circle and enable individuals to step out of the ground and take the first step. The whole world needs this project. In metropolitan cities, there are people who can find these opportunities by forcing them. But in the remaining cities of the world, which are more numerous, there are billions of individuals who cannot access even one of these opportunities. This project is applicable in all cities of the world. I can also take part in this project.